requirements.txt
asgiref==3.5.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
cachetools==5.2.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
certifi==2022.9.24 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
charset-normalizer==2.1.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
django-model-utils==4.2.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
django==4.1.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
djangorestframework==3.14.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
idna==3.4 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
markdown==3.4.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
python-dateutil==2.8.2 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
pytz==2022.5 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
requests==2.28.1 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
six==1.16.0 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
sqlparse==0.4.3 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
tzdata==2022.5 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0" and sys_platform == "win32"
urllib3==1.26.12 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
zenpy==2.0.25 ; python_full_version >= "3.10.0" and python_full_version < "3.11.0"
